``warn_until`` now handles instances of ``SaltVersion``. Migrate tests to Pytest.
