+++ salt/utils/versions.py
@@ -110,14 +110,26 @@ def warn_until(
                                 issued. When we're only after the salt version
                                 checks to raise a ``RuntimeError``.
     """
-    if not isinstance(version, (tuple, str, salt.version.SaltStackVersion)):
+    if not isinstance(
+        version, (tuple, str, salt.version.SaltVersion, salt.version.SaltStackVersion)
+    ):
         raise RuntimeError(
             "The 'version' argument should be passed as a tuple, string or "
-            "an instance of 'salt.version.SaltStackVersion'."
+            "an instance of 'salt.version.SaltVersion' or "
+            "'salt.version.SaltStackVersion'."
         )
+    elif isinstance(version, salt.version.SaltVersion):
+        version = salt.version.SaltStackVersion(*version.info)
     elif isinstance(version, tuple):
         version = salt.version.SaltStackVersion(*version)
     elif isinstance(version, str):
+        if version.lower() not in salt.version.SaltStackVersion.LNAMES:
+            raise RuntimeError(
+                "Incorrect spelling for the release name in the warn_utils "
+                "call. Expecting one of these release names: {}".format(
+                    [vs.name for vs in salt.version.SaltVersionsInfo.versions()]
+                )
+            )
         version = salt.version.SaltStackVersion.from_name(version)
 
     if stacklevel is None:
